redis mal

+++ python-flask-restapi/app.py
@@ -1,6 +1,6 @@
 import json
 import os
-
+import redis
 from app_service import AppService
 from db import Database
 from dbMongo import MongoDatabase  # Importa la clase MongoDatabase
@@ -14,7 +14,6 @@ from flask_jwt_extended import (
     create_access_token,
     get_jwt_identity,
     verify_jwt_in_request,
-    get_jwt,
 )
 
 
@@ -25,23 +24,36 @@ DB_NAME = os.getenv("DB_NAME_POSTGRES")
 DB_USER = os.getenv("DB_USER_POSTGRES")
 DB_PASSWORD = os.getenv("DB_PASSWORD_POSTGRES")
 
-# Configuración de la base de datos MongoDB
-MONGO_HOST = os.getenv("MONGO_HOST")
-MONGO_PORT = os.getenv("MONGO_PORT")
-
 # Inicializar la conexión a la base de datos PostgreSQL
 db = Database(
     database=DB_NAME, host=DB_HOST, user=DB_USER, password=DB_PASSWORD, port=DB_PORT
 )
 
+# Configuración de la base de datos MongoDB
+MONGO_HOST = os.getenv("DB_HOST_MONGO")
+MONGO_PORT = os.getenv("DB_PORT_MONGO")
+MONGO_USER = os.getenv("DB_USER_MONGO")
+MONGO_PASSWORD = os.getenv("DB_PASSWORD_MONGO")
+
 # Inicializar la conexión a la base de datos MongoDB
-client = MongoClient("mongo", 27017, username="root", password="password")
+client = MongoClient(
+    f"mongodb://{MONGO_USER}:{MONGO_PASSWORD}@{MONGO_HOST}:{MONGO_PORT}/?authSource=admin"
+)
 mongo_db = MongoDatabase(client)
 
-# Inicializar la instancia de AppService con ambas conexiones de base de datos
-appService = AppService(db, mongo_db)
+# Configuración de la base de datos Redis
+REDIS_HOST = os.getenv("REDIS_HOST")
+REDIS_PORT = os.getenv("REDIS_PORT")
+REDIS_DB = os.getenv("REDIS_DB")
 
-# Inicializar la aplicación Flask
+# Inicializar la conexión a la base de datos Redis
+redis_client = redis.StrictRedis(host=REDIS_HOST, port=REDIS_PORT, db=REDIS_DB)
+
+# Inicializar la instancia de AppService con ambas conexiones de base de datos
+appService = AppService(db, mongo_db, redis_client)
+
+
+# ------------------------------------------------------------------- Inicializar la aplicación Flask -------------------------------------------------------------------
 app = Flask(__name__)
 app.config["JWT_SECRET_KEY"] = "1234"  # Cambia esto por tu clave secreta
 jwt = JWTManager(app)
@@ -106,8 +118,9 @@ def check_role(required_role):
         return False
 
 
-# --------------------------------------------------------------------------   Usuarios   --------------------------------------------------------------------------
+# --------------------------------------------------------------------------   USUARIOS   --------------------------------------------------------------------------
 @app.route("/users")
+@jwt_required()
 def users():
     if not check_role(1):
         return (
@@ -123,6 +136,7 @@ def user_by_id(id):
 
 
 @app.route("/users/<int:id>", methods=["PUT"])
+@jwt_required()
 def update_user(id):
     verify_jwt_in_request()
     # Obtener el ID del usuario actual desde el token JWT
@@ -138,14 +152,20 @@ def update_user(id):
 
 
 @app.route("/users/<int:id>", methods=["DELETE"])
+@jwt_required()  # Requiere autenticación JWT
 def delete_user(id):
+    if not check_role(1):
+        return (
+            jsonify({"error": "Usuario no autorizado para realizar esta acción"}),
+            403,
+        )
     return appService.delete_user(str(id))
 
 
-# --------------------------------------------------------------------------   Encuestas    --------------------------------------------------------------------------
+# --------------------------------------------------------------------------   ENCUESTAS    --------------------------------------------------------------------------
 @app.route("/surveys")
 def encuestas():
-    return appService.get_encuestas()
+    return appService.get_encuestas()  # solo las que tengan publicadas
 
 
 @app.route("/surveys/<int:id>")
@@ -157,7 +177,10 @@ def encuesta_by_id(id):
 @jwt_required()  # Requiere autenticación JWT
 def create_encuesta():
     if check_role(3):
-        return jsonify({"error": "Usuario no autorizado para crear encuestas"}), 403
+        return (
+            jsonify({"error": "Usuario no autorizado para crear encuestas"}),
+            403,
+        )  # si no es admin ni creador no puede crear encuestas
 
     request_data = request.get_json()
     encuesta = request_data
@@ -173,7 +196,12 @@ def create_encuesta():
 
 
 @app.route("/surveys/<int:id>", methods=["PUT"])
+@jwt_required()  # Requiere autenticación JWT
 def update_encuesta(id):
+
+    if check_role(3):
+        return jsonify({"error": "Usuario no autorizado para crear encuestas"}), 403
+
     request_data = request.get_json()
     result = appService.update_encuesta(request_data, str(id))
     # Verifica si la creación de la encuesta fue exitosa
@@ -184,13 +212,14 @@ def update_encuesta(id):
 
 
 @app.route("/surveys/<int:id>", methods=["DELETE"])
+@jwt_required()  # Requiere autenticación JWT
 def delete_encuesta(id):
+    if check_role(3):
+        return (
+            jsonify({"error": "Usuario no autorizado para crear encuestas"}),
+            403,
+        )  # si no es admin ni creador no puede borrar encuestas
     return appService.delete_encuesta(str(id))
-
-
-@app.route("/encuestassql")
-def encuestassql():
-    return appService.get_encuestassql()
 
 
 @app.route("/surveys/<int:id>/publish", methods=["POST"])
@@ -212,13 +241,17 @@ def publish_survey(id):
         return jsonify({"error": "Error al publicar la encuesta"}), 500
 
 
-# --------------------------------------------------------------- Preguntas de Encuestas ---------------------------------------------------------------
+# --------------------------------------------------------------- PREGUNTAS DE ENCUESTAS ---------------------------------------------------------------
 
 
 @app.route("/surveys/<int:id>/questions", methods=["POST"])
 @jwt_required()  # Requiere autenticación JWT
 def add_question(id):
-
+    if check_role(3):
+        return (
+            jsonify({"error": "Usuario no autorizado para crear encuestas"}),
+            403,
+        )  # si no es admin ni creador no puede
     request_data = request.get_json()
     # Llama al método del servicio para agregar la pregunta a la encuesta
     result = appService.add_question(str(id), request_data)
@@ -239,7 +272,11 @@ def get_questions(id):
 @app.route("/surveys/<int:id>/questions/<questionId>", methods=["PUT"])
 @jwt_required()  # Requiere autenticación JWT
 def update_question(id, questionId):
-
+    if check_role(3):
+        return (
+            jsonify({"error": "Usuario no autorizado para crear encuestas"}),
+            403,
+        )  # si no es admin ni creador no puede
     request_data = request.get_json()
     # Llama al método del servicio para actualizar la pregunta de la encuesta
     result = appService.update_question(str(id), int(questionId), request_data)
@@ -253,7 +290,11 @@ def update_question(id, questionId):
 @app.route("/surveys/<int:id>/questions/<questionId>", methods=["DELETE"])
 @jwt_required()  # Requiere autenticación JWT
 def delete_question(id, questionId):
-
+    if check_role(3):
+        return (
+            jsonify({"error": "Usuario no autorizado para crear encuestas"}),
+            403,
+        )  # si no es admin ni creador no
     # Llama al método del servicio para eliminar la pregunta de la encuesta
     result = appService.delete_question(str(id), int(questionId))
 
@@ -261,6 +302,60 @@ def delete_question(id, questionId):
         return jsonify({"message": "Pregunta eliminada correctamente"}), 200
     else:
         return jsonify({"error": "Error al eliminar la pregunta"}), 500
+
+
+# ------------------------------------------------------------- RESPUESTAS -------------------------------------------------------------
+
+
+@app.route("/surveys/<string:id>/responses", methods=["POST"])
+def submit_response(id):
+    try:
+        respuesta = request.get_json()
+        encuesta_id = str(id)
+        usuario_id = respuesta["usuario_id"]
+        response_data = respuesta["respuestas"]
+
+        if not encuesta_id or not usuario_id or not response_data:
+            return jsonify({"error": "Datos incompletos"}), 400
+
+        result = appService.submit_response(encuesta_id, usuario_id, response_data)
+
+        if result:
+            return jsonify({"message": "Respuesta enviada correctamente"}), 201
+        else:
+            return jsonify({"error": "Error al enviar la respuesta"}), 500
+
+    except Exception as e:
+        print(f"Error en la solicitud POST /surveys/{id}/responses: {str(e)}")
+        return jsonify({"error": "Error interno del servidor"}), 500
+
+
+@app.route("/surveys/<string:id>/responses", methods=["GET"])
+@jwt_required()  # Requiere autenticación JWT
+def get_responses(id):
+
+    if check_role(3):
+        return (
+            jsonify({"error": "Usuario no autorizado para crear encuestas"}),
+            403,
+        )  # si no es admin ni creador no puede
+
+    try:
+        encuesta_id = str(id)
+
+        if not encuesta_id:
+            return jsonify({"error": "ID de encuesta faltante"}), 400
+
+        responses = appService.get_responses(encuesta_id)
+
+        if responses is not None:
+            return jsonify(responses), 200
+        else:
+            return jsonify({"error": "Error al obtener las respuestas"}), 500
+
+    except Exception as e:
+        print(f"Error en la solicitud GET /surveys/{id}/responses: {str(e)}")
+        return jsonify({"error": "Error interno del servidor"}), 500
 
 
 # ------------------------------------------------------------- ENCUESTADOS -------------------------------------------------------------
@@ -304,3 +399,32 @@ def update_respondent(id):
 @app.route("/respondents/<int:id>", methods=["DELETE"])
 def delete_respondent(id):
     return appService.delete_respondent(str(id))
+
+
+# ------------------------------------------------------------- ANÁLISIS DE ENCUESTAS -------------------------------------------------------------
+@app.route("/surveys/<string:id>/analysis", methods=["GET"])
+@jwt_required()  # Requiere autenticación JWT
+def generate_analysis(id):
+
+    if check_role(3):
+        return (
+            jsonify({"error": "Usuario no autorizado para crear encuestas"}),
+            403,
+        )  # si no es admin ni creador no puede crear encuestas
+
+    try:
+        encuesta_id = str(id)
+
+        if not encuesta_id:
+            return jsonify({"error": "ID de encuesta faltante"}), 400
+
+        success, analysis = appService.generate_analysis(encuesta_id)
+
+        if success:
+            return jsonify({"analysis": analysis}), 200
+        else:
+            return jsonify({"error": "Error al generar el análisis"}), 400
+
+    except Exception as e:
+        print(f"Error en la solicitud GET /surveys/{id}/analysis: {str(e)}")
+        return jsonify({"error": "Error interno del servidor"}), 500
